requirements.txt
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.11
click==6.6
itsdangerous==0.24
pycodestyle==2.2.0
python-dotenv==0.6.3
